add new pizza and cloudinary functions added

+++ pizzaapp/pages/api/products/[id].js
@@ -2,7 +2,9 @@ import dbConnect from '../../../util/mongo'
 import Product from '../../../models/Product'
 
 export default async function handler(req, res){
-    const { method, query:{id}} = req
+    const { method, cookies, query:{id}} = req
+
+    const token = cookies.token
 
     dbConnect()
 
@@ -16,6 +18,10 @@ export default async function handler(req, res){
         }
      }
     if (method==='PUT') {
+        if(!token || token !== process.env.token){
+            return res.status(401).json('Not authenticated')
+        }
+
         try {
             const product = await Product.create(req.body)
             res.status(201).json(product)
@@ -24,6 +30,10 @@ export default async function handler(req, res){
         }
     }
     if (method==='DELETE') {
+        if(!token || token !== process.env.token){
+            return res.status(401).json('Not authenticated')
+        }
+
         try {
             await Product.findByIdAndDelete(id)
             res.status(200).json('The product has been deleted')
